Allow downloading legacy files.
